+++ src/contents/ui/ExchangeModel.qml
@@ -8,17 +8,7 @@
  */
 
 import QtQuick 2.0
-import org.kde.plasma.configuration 2.0
 
-ConfigModel {
-	ConfigCategory {
-		name: i18n("Exchanges")
-		icon: "view-visible"
-		source: "configExchanges.qml"
-	}
-	ConfigCategory {
-		name: i18n("Layout")
-		icon: "view-visible"
-		source: "configLayout.qml"
-	}
+ListModel {
+    id: exchangeModel
 }