+++ CHANGES.md
@@ -1,6 +1,7 @@
 * dev
+  * Reworked exchange management and added support for unlimited number of exchanges.
+  * Exchanges can be now easily reordered.
   * Added support for Plasma 5.19+ widget background controls.
-
 
 * v1.2.0 (2021-03-24)
   * Internal widget layout grid is configurable now. Requested by @Foul [#11]